% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regressionline_function.R
\name{gettingmean}
\alias{gettingmean}
\title{Calculate mean}
\usage{
gettingmean(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
The mean of \code{x}, omitting any missing values
}
\description{
Calculate mean dropping any NAs
}
\examples{
mean(c(1,NA,5,6))
}

