% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regressionline_function.R
\name{Regressionline_function}
\alias{Regressionline_function}
\title{A Regressionline Function}
\usage{
Regressionline_function(numb = 100, mu_x = 10, mu_y = 20)
}
\arguments{
\item{numb}{How many points you want to generate? Defaults 100}

\item{mu_x}{The mean of normal distribution that generates x Defaults 10}

\item{mu_y}{The mean of normal distribution that generates y Defaults 20}
}
\value{
None
}
\description{
This function allows you to plot a regression line based on points generated
from two normal distribution
}
\examples{
Regressionline_function()
}

